+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 ### Doing
 
 ### Done
-- 2025-04-11 : tests: improvements
+- 2025-04-11 : 
+    - tests: improvements
+    - tests: refacto
 - 2025-04-10 : tests: improve frontend tests with Playwright workflows
 - 2025-04-09 : tests: improving frontend tests with Playwright
 - 2025-04-08 : chore: local css in dedicated files